Retryable: time given as long instead of int to play better with TimeInterval

+++ net/src/main/java/com/elastisys/scale/commons/net/retryable/DelayStrategy.java
@@ -16,5 +16,5 @@ public interface DelayStrategy {
      * @param elapsedTimeMillis
      *            The elapsed time in milliseconds since the first attempt.
      */
-    void introduceDelay(int failedAttempts, long elapsedTimeMillis);
+    void introduceDelay(long failedAttempts, long elapsedTimeMillis);
 }
